libext2: avoid use of __divdi3()/__moddi3()
Changing the type of offset to off_t has implictely pulled
the use of __divdi3()//__moddi3() that we are not importing from
the system. Cast offset to int to avoid this

+++ libext2/ext2_utils.c
@@ -244,8 +244,9 @@ int ext2_read_data(ext2_VOLUME* volume, struct ext2_inode *inode,
 		length = inode->i_size - offset;
 
 	read = 0;
-	logical = offset / blocksize;
-	shift = offset % blocksize;
+	/* cast to int to avoid use of divdi3 and moddi3 */
+	logical = (int)offset / blocksize;
+	shift = (int)offset % blocksize;
 
 	if (shift) {
 		physical = ext2_get_block_addr(volume, inode, logical);
